Refactor: Entity 패키지, Model 패키지 구분
- 객체-관계 간 매핑을 위한 객체는 entities 패키지에 정의
- 순수 비즈니스 로직과 관련된 객체는 models 패키지에 정의
