+++ src/main/java/com/wanted/jaringoby/common/filters/AuthorizationFilter.java
@@ -1,8 +1,8 @@
 package com.wanted.jaringoby.common.filters;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.wanted.jaringoby.domains.customer.entities.CustomerId;
 import com.wanted.jaringoby.domains.customer.exceptions.CustomerNotFoundException;
-import com.wanted.jaringoby.domains.customer.models.customer.CustomerId;
 import com.wanted.jaringoby.domains.customer.repositories.CustomerRepository;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
