+++ src/main/java/com/wanted/jaringoby/common/validations/groups/PatternMatchesGroup.java
@@ -1,5 +1,6 @@
 package com.wanted.jaringoby.common.validations.groups;
 
+// TODO: PatternMatchGroup으로 이름 변경
 public interface PatternMatchesGroup {
 
 }
