+++ src/main/java/com/wanted/jaringoby/domains/customer/applications/CreateCustomerService.java
@@ -3,9 +3,9 @@ package com.wanted.jaringoby.domains.customer.applications;
 import com.wanted.jaringoby.common.utils.UlidGenerator;
 import com.wanted.jaringoby.domains.customer.dtos.CreateCustomerRequestDto;
 import com.wanted.jaringoby.domains.customer.dtos.CreateCustomerResponseDto;
+import com.wanted.jaringoby.domains.customer.entities.Customer;
 import com.wanted.jaringoby.domains.customer.exceptions.CustomerReconfirmPasswordMismatchedException;
 import com.wanted.jaringoby.domains.customer.exceptions.CustomerUsernameDuplicatedException;
-import com.wanted.jaringoby.domains.customer.models.customer.Customer;
 import com.wanted.jaringoby.domains.customer.repositories.CustomerRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
