+++ src/main/java/com/wanted/jaringoby/domains/expense/repositories/ExpenseRepository.java
@@ -1,7 +1,7 @@
 package com.wanted.jaringoby.domains.expense.repositories;
 
-import com.wanted.jaringoby.domains.expense.models.Expense;
-import com.wanted.jaringoby.domains.expense.models.ExpenseId;
+import com.wanted.jaringoby.domains.expense.entities.Expense;
+import com.wanted.jaringoby.domains.expense.entities.ExpenseId;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ExpenseRepository extends JpaRepository<Expense, ExpenseId> {
