+++ src/main/java/com/wanted/jaringoby/domains/ledger/applications/BudgetRecommendationService.java
@@ -6,6 +6,8 @@ import com.wanted.jaringoby.domains.category.models.Category;
 import com.wanted.jaringoby.domains.ledger.dtos.GetBudgetRecommendationQueryParamsDto;
 import com.wanted.jaringoby.domains.ledger.dtos.GetBudgetRecommendationResponseDto;
 import com.wanted.jaringoby.domains.ledger.dtos.query.LedgerIdAndBudgetsQueryResultDto;
+import com.wanted.jaringoby.domains.ledger.entities.budget.Budget;
+import com.wanted.jaringoby.domains.ledger.entities.ledger.LedgerId;
 import com.wanted.jaringoby.domains.ledger.exceptions.CategoryCannotBeIncludedException;
 import com.wanted.jaringoby.domains.ledger.exceptions.TruncationScaleGreaterThanTotalAmountException;
 import com.wanted.jaringoby.domains.ledger.exceptions.TruncationScaleIndivisibleException;
@@ -13,8 +15,6 @@ import com.wanted.jaringoby.domains.ledger.exceptions.TruncationScaleRangeNotAll
 import com.wanted.jaringoby.domains.ledger.models.BudgetAmountCalculator;
 import com.wanted.jaringoby.domains.ledger.models.BudgetAmountRatioCalculator;
 import com.wanted.jaringoby.domains.ledger.models.BudgetAmountTruncator;
-import com.wanted.jaringoby.domains.ledger.models.budget.Budget;
-import com.wanted.jaringoby.domains.ledger.models.ledger.LedgerId;
 import com.wanted.jaringoby.domains.ledger.repositories.BudgetRepository;
 import java.util.List;
 import java.util.Map;
