+++ src/main/java/com/wanted/jaringoby/domains/ledger/applications/CreateLedgerService.java
@@ -6,14 +6,14 @@ import com.wanted.jaringoby.common.utils.UlidGenerator;
 import com.wanted.jaringoby.domains.category.exceptions.CategoryDuplicatedException;
 import com.wanted.jaringoby.domains.category.exceptions.CategoryNotFoundException;
 import com.wanted.jaringoby.domains.category.models.Category;
-import com.wanted.jaringoby.domains.customer.models.customer.CustomerId;
+import com.wanted.jaringoby.domains.customer.entities.CustomerId;
 import com.wanted.jaringoby.domains.ledger.dtos.BudgetRequestDto;
 import com.wanted.jaringoby.domains.ledger.dtos.CreateLedgerRequestDto;
 import com.wanted.jaringoby.domains.ledger.dtos.CreateLedgerResponseDto;
+import com.wanted.jaringoby.domains.ledger.entities.budget.Budget;
+import com.wanted.jaringoby.domains.ledger.entities.ledger.Ledger;
 import com.wanted.jaringoby.domains.ledger.exceptions.LedgerPeriodInvalidException;
 import com.wanted.jaringoby.domains.ledger.exceptions.LedgerPeriodOverlappedException;
-import com.wanted.jaringoby.domains.ledger.models.budget.Budget;
-import com.wanted.jaringoby.domains.ledger.models.ledger.Ledger;
 import com.wanted.jaringoby.domains.ledger.repositories.BudgetRepository;
 import com.wanted.jaringoby.domains.ledger.repositories.LedgerRepository;
 import java.time.LocalDate;
