+++ src/main/java/com/wanted/jaringoby/domains/ledger/models/BudgetAmountRatioCalculator.java
@@ -3,8 +3,8 @@ package com.wanted.jaringoby.domains.ledger.models;
 import com.wanted.jaringoby.common.models.Money;
 import com.wanted.jaringoby.common.models.Percentage;
 import com.wanted.jaringoby.domains.category.models.Category;
-import com.wanted.jaringoby.domains.ledger.models.budget.Budget;
-import com.wanted.jaringoby.domains.ledger.models.ledger.LedgerId;
+import com.wanted.jaringoby.domains.ledger.entities.budget.Budget;
+import com.wanted.jaringoby.domains.ledger.entities.ledger.LedgerId;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
