+++ src/main/java/com/wanted/jaringoby/domains/ledger/repositories/BudgetQueryDslRepositoryImpl.java
@@ -6,8 +6,8 @@ import static com.wanted.jaringoby.domains.ledger.models.ledger.QLedger.ledger;
 import com.querydsl.core.Tuple;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import com.wanted.jaringoby.domains.ledger.dtos.query.LedgerIdAndBudgetsQueryResultDto;
-import com.wanted.jaringoby.domains.ledger.models.budget.Budget;
-import com.wanted.jaringoby.domains.ledger.models.ledger.LedgerId;
+import com.wanted.jaringoby.domains.ledger.entities.budget.Budget;
+import com.wanted.jaringoby.domains.ledger.entities.ledger.LedgerId;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
