+++ src/main/java/com/wanted/jaringoby/domains/ledger/repositories/LedgerRepository.java
@@ -1,7 +1,7 @@
 package com.wanted.jaringoby.domains.ledger.repositories;
 
-import com.wanted.jaringoby.domains.ledger.models.ledger.Ledger;
-import com.wanted.jaringoby.domains.ledger.models.ledger.LedgerId;
+import com.wanted.jaringoby.domains.ledger.entities.ledger.Ledger;
+import com.wanted.jaringoby.domains.ledger.entities.ledger.LedgerId;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LedgerRepository extends JpaRepository<Ledger, LedgerId>,
