+++ src/main/java/com/wanted/jaringoby/session/applications/LogoutService.java
@@ -1,6 +1,6 @@
 package com.wanted.jaringoby.session.applications;
 
-import com.wanted.jaringoby.domains.customer.models.customer.CustomerId;
+import com.wanted.jaringoby.domains.customer.entities.CustomerId;
 import com.wanted.jaringoby.session.repositories.CustomerRefreshTokenRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
