+++ src/test/java/com/wanted/jaringoby/domains/customer/applications/CreateCustomerServiceTest.java
@@ -11,9 +11,9 @@ import static org.mockito.Mockito.verify;
 import com.wanted.jaringoby.common.utils.UlidGenerator;
 import com.wanted.jaringoby.domains.customer.dtos.CreateCustomerRequestDto;
 import com.wanted.jaringoby.domains.customer.dtos.CreateCustomerResponseDto;
+import com.wanted.jaringoby.domains.customer.entities.Customer;
 import com.wanted.jaringoby.domains.customer.exceptions.CustomerReconfirmPasswordMismatchedException;
 import com.wanted.jaringoby.domains.customer.exceptions.CustomerUsernameDuplicatedException;
-import com.wanted.jaringoby.domains.customer.models.customer.Customer;
 import com.wanted.jaringoby.domains.customer.repositories.CustomerRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
