+++ src/test/java/com/wanted/jaringoby/domains/ledger/applications/ModifyLedgerPeriodServiceTest.java
@@ -8,15 +8,15 @@ import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.verify;
 
 import com.wanted.jaringoby.common.constants.Date;
-import com.wanted.jaringoby.domains.customer.models.customer.CustomerId;
+import com.wanted.jaringoby.domains.customer.entities.CustomerId;
 import com.wanted.jaringoby.domains.ledger.dtos.ModifyLedgerPeriodRequestDto;
+import com.wanted.jaringoby.domains.ledger.entities.ledger.Ledger;
+import com.wanted.jaringoby.domains.ledger.entities.ledger.LedgerId;
 import com.wanted.jaringoby.domains.ledger.exceptions.LedgerNotFoundException;
 import com.wanted.jaringoby.domains.ledger.exceptions.LedgerNotOwnedException;
 import com.wanted.jaringoby.domains.ledger.exceptions.LedgerPeriodEndedException;
 import com.wanted.jaringoby.domains.ledger.exceptions.LedgerPeriodInvalidException;
 import com.wanted.jaringoby.domains.ledger.exceptions.LedgerPeriodOverlappedException;
-import com.wanted.jaringoby.domains.ledger.models.ledger.Ledger;
-import com.wanted.jaringoby.domains.ledger.models.ledger.LedgerId;
 import com.wanted.jaringoby.domains.ledger.repositories.LedgerRepository;
 import java.time.LocalDate;
 import java.util.Optional;
@@ -309,7 +309,7 @@ class ModifyLedgerPeriodServiceTest {
             }
         }
 
-        @DisplayName("다른 고객의 예산 관리인 경우 경우")
+        @DisplayName("다른 고객의 예산 관리인 경우")
         @Nested
         class LedgerNotOwned {
