+++ src/test/java/com/wanted/jaringoby/domains/ledger/controllers/LedgerControllerTest.java
@@ -17,7 +17,7 @@ import com.wanted.jaringoby.common.config.security.SecurityConfig;
 import com.wanted.jaringoby.common.config.validation.ValidationConfig;
 import com.wanted.jaringoby.common.utils.JwtUtil;
 import com.wanted.jaringoby.common.validations.BindingResultChecker;
-import com.wanted.jaringoby.domains.customer.models.customer.CustomerId;
+import com.wanted.jaringoby.domains.customer.entities.CustomerId;
 import com.wanted.jaringoby.domains.customer.repositories.CustomerRepository;
 import com.wanted.jaringoby.domains.ledger.applications.BudgetRecommendationService;
 import com.wanted.jaringoby.domains.ledger.applications.CreateLedgerService;
