+++ src/test/java/com/wanted/jaringoby/domains/ledger/models/BudgetAmountRatioCalculatorTest.java
@@ -5,8 +5,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import com.wanted.jaringoby.common.models.Money;
 import com.wanted.jaringoby.common.models.Percentage;
 import com.wanted.jaringoby.domains.category.models.Category;
-import com.wanted.jaringoby.domains.ledger.models.budget.Budget;
-import com.wanted.jaringoby.domains.ledger.models.ledger.LedgerId;
+import com.wanted.jaringoby.domains.ledger.entities.budget.Budget;
+import com.wanted.jaringoby.domains.ledger.entities.ledger.LedgerId;
 import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
