+++ src/test/java/com/wanted/jaringoby/domains/ledger/repositories/BudgetQueryDslRepositoryImplTest.java
@@ -6,8 +6,8 @@ import com.wanted.jaringoby.common.constants.DateTime;
 import com.wanted.jaringoby.config.jpa.JpaTestConfig;
 import com.wanted.jaringoby.domains.category.models.Category;
 import com.wanted.jaringoby.domains.ledger.dtos.query.LedgerIdAndBudgetsQueryResultDto;
-import com.wanted.jaringoby.domains.ledger.models.budget.Budget;
-import com.wanted.jaringoby.domains.ledger.models.ledger.LedgerId;
+import com.wanted.jaringoby.domains.ledger.entities.budget.Budget;
+import com.wanted.jaringoby.domains.ledger.entities.ledger.LedgerId;
 import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
