+++ src/test/java/com/wanted/jaringoby/domains/ledger/repositories/LedgerQueryDslRepositoryImplTest.java
@@ -5,9 +5,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import com.wanted.jaringoby.common.constants.Date;
 import com.wanted.jaringoby.common.constants.DateTime;
 import com.wanted.jaringoby.config.jpa.JpaTestConfig;
-import com.wanted.jaringoby.domains.customer.models.customer.CustomerId;
-import com.wanted.jaringoby.domains.ledger.models.ledger.Ledger;
-import com.wanted.jaringoby.domains.ledger.models.ledger.LedgerId;
+import com.wanted.jaringoby.domains.customer.entities.CustomerId;
+import com.wanted.jaringoby.domains.ledger.entities.ledger.Ledger;
+import com.wanted.jaringoby.domains.ledger.entities.ledger.LedgerId;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
