+++ src/test/java/com/wanted/jaringoby/session/applications/ReissueAccessTokenServiceTest.java
@@ -6,7 +6,7 @@ import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
 
 import com.wanted.jaringoby.common.utils.JwtUtil;
-import com.wanted.jaringoby.domains.customer.models.customer.CustomerId;
+import com.wanted.jaringoby.domains.customer.entities.CustomerId;
 import com.wanted.jaringoby.session.dtos.ReissueAccessTokenResponseDto;
 import com.wanted.jaringoby.session.exceptions.CustomerRefreshTokenNotFoundException;
 import com.wanted.jaringoby.session.repositories.CustomerRefreshTokenRepository;
